requirement.txt
# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Deep Learning & SMPL-X
torch>=1.9.0
smplx>=0.1.28

# MuJoCo Physics Engine
mujoco>=3.0.0

# PyQt5 GUI Framework
PyQt5>=5.15.0

# PyQt5 Additional Components
pyqtgraph>=0.12.0
pyqdarktheme>=2.0.0
